Modifications on Stripe methods location
