+++ src/main/java/com/kuna_backend/controllers/OrderController.java
@@ -1,8 +1,6 @@
 package com.kuna_backend.controllers;
 
 import com.kuna_backend.common.ApiResponse;
-import com.kuna_backend.dtos.checkout.CheckoutItemDto;
-import com.kuna_backend.dtos.checkout.StripeResponse;
 import com.kuna_backend.exceptions.AuthenticationFailException;
 import com.kuna_backend.exceptions.OrderNotFoundException;
 import com.kuna_backend.models.Client;
@@ -11,8 +9,6 @@ import com.kuna_backend.services.AuthenticationService;
 import com.kuna_backend.services.ItemService;
 import com.kuna_backend.services.OrderService;
 
-import com.stripe.exception.StripeException;
-import com.stripe.model.checkout.Session;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -32,15 +28,6 @@ public class OrderController {
     @Autowired
     private ItemService itemService;
 
-    // Stripe Create Session Api
-    @PostMapping("/create-checkout-session")
-    public ResponseEntity<StripeResponse> checkoutList(@RequestBody List<CheckoutItemDto> checkoutItemDtoList) throws StripeException {
-        // Create the Stripe session
-        Session session = orderService.createSession(checkoutItemDtoList);
-        StripeResponse stripeResponse = new StripeResponse(session.getId());
-        // Send the Stripe session id in response
-        return new ResponseEntity<StripeResponse>(stripeResponse, HttpStatus.OK);
-    }
 
     // POST Endpoint - Place order after Checkout session
     @PostMapping("/add")
@@ -67,7 +54,7 @@ public class OrderController {
         return new ResponseEntity<>(orderDtoList, HttpStatus.OK);
     }
 
-    // GET Order Items for an Order
+    // GET Order Items for an Order / Endpoint
     @GetMapping(path = "/{id}")
     public ResponseEntity<Object> getOrderById(@PathVariable ("id") Integer id, @RequestParam ("token") String token)
             throws AuthenticationFailException {
@@ -82,38 +69,11 @@ public class OrderController {
         }
     }
 
-
-    // UPDATE an Order / Endpoint
-//    @PutMapping(path = "/{id}")
-//    public ResponseEntity<Order> update(@RequestBody Order order, @PathVariable Integer id) {
-//        try {
-//            Order existOrder = orderService.getOrder(id);
-//            order.setId(id);
-//            orderService.createOrder(order);
-//            return new ResponseEntity<>(HttpStatus.OK);
-//        } catch (NoSuchElementException e) {
-//            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-//        }
-//    }
-
     //DELETE one Order by ID / Endpoint
     @DeleteMapping(path = "/{id}")
     public void delete (@PathVariable Integer id) {
         orderService.deleteOrder(id);
     }
-
-
-//    // Retrieve all Items for an Order / Many-to-many relationship Endpoint
-//    @GetMapping(path = "/{id}/items")
-//    public Set<Item> retrieveItemsForOrder(@PathVariable Integer id) {
-//
-//        Order order = orderService.getOrder(id);
-//
-//        if (order==null)
-//            throw new NoSuchElementException("id:"+id);
-//
-//        return order.getItems();
-//    }
 
 }
